requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
mysqlclient==2.1.0
bcrypt==3.2.0
PyJWT==2.3.0